// interfaces/character.tsx
export interface Origin {
    name: string;
    url: string;
    speciesData: Map<string,number>;
  }
  
  interface Location {
    name: string;
    url: string;
  }
  
  interface Character {
    id: number;
    name: string;
    status: string;
    species: string;
    type: string;
    gender: string;
    origin: Origin;
    location: Location;
    image: string;
    episode: string[];
    url: string;
    created: string;
  }
  
  export default Character;